update route tests
