+++ src/test/features/ccj/routes/confirmation-ccj.ts
@@ -1,0 +1,57 @@
+import { expect } from 'chai'
+import * as request from 'supertest'
+import * as config from 'config'
+
+import { attachDefaultHooks } from 'test/routes/hooks'
+import 'test/routes/expectations'
+import { checkAuthorizationGuards } from 'test/features/ccj/routes/checks/authorization-check'
+
+import { Paths as CCJPaths } from 'ccj/paths'
+
+import { app } from 'main/app'
+
+import * as idamServiceMock from 'test/http-mocks/idam'
+import * as claimStoreServiceMock from 'test/http-mocks/claim-store'
+import { checkNotClaimantInCaseGuard } from 'test/features/ccj/routes/checks/not-claimant-in-case-check'
+
+const externalId = claimStoreServiceMock.sampleClaimObj.externalId
+const cookieName: string = config.get<string>('session.cookieName')
+const pagePath = CCJPaths.ccjConfirmationPage.evaluateUri({ externalId: externalId })
+
+describe('CCJ: confirmation page', () => {
+  attachDefaultHooks(app)
+
+  describe('on GET', () => {
+    const method = 'get'
+    checkAuthorizationGuards(app, method, pagePath)
+    checkNotClaimantInCaseGuard(app, method, pagePath)
+
+    describe('for authorized user', () => {
+      beforeEach(() => {
+        idamServiceMock.resolveRetrieveUserFor('1', 'citizen')
+      })
+
+      context('when user authorised', () => {
+        it('should return 500 and render error page when cannot retrieve claims', async () => {
+          claimStoreServiceMock.rejectRetrieveClaimByExternalId('HTTP error')
+
+          await request(app)
+            .get(pagePath)
+            .set('Cookie', `${cookieName}=ABC`)
+            .expect(res => expect(res).to.be.serverError.withText('Error'))
+        })
+
+        it('should render page when everything is fine', async () => {
+          claimStoreServiceMock.resolveRetrieveClaimByExternalId(
+            { countyCourtJudgmentRequestedAt: '2017-10-10T22:45:51.785' }
+          )
+
+          await request(app)
+            .get(pagePath)
+            .set('Cookie', `${cookieName}=ABC`)
+            .expect(res => expect(res).to.be.successful.withText('County Court Judgment requested'))
+        })
+      })
+    })
+  })
+})
